Forked for Pimoroni interstate 75W board

+++ setup/frames.py
@@ -1,3 +1,5 @@
-# Loop setup
-PERIOD = 0.1
-PER_SECOND = 1 / PERIOD
+# Animation timing constants for Interstate 75 W
+# Loop setup - timing in seconds
+
+PERIOD = 0.1  # 100ms per frame (10 FPS)
+PER_SECOND = int(1 / PERIOD)  # 10 frames per second
